fix attandance calculation

+++ src/main/java/com/example/leavemanagement/dto/AttendanceReportSummary.java
@@ -10,6 +10,8 @@ package com.example.leavemanagement.dto;
  * @param projectId the project this summary is scoped to (may be null if the resource has no
  *     assignment)
  * @param period human-readable period label, e.g. "July 2026", "Q3 2026", "2026"
+ * @param presentDays effective days worked: P-status days + halfDays×0.5 (worked portion of
+ *     half-days). Formula: workingDays − (absentDays + halfDays×0.5).
  * @param paidLeaveDays effective absences within the period's paid-leave entitlement:
  *     MIN(absentDays + halfDays×0.5, leaveLimit). Each half-day consumes 0.5 leave days.
  *     Zero when no leave policy is found.
@@ -22,7 +24,7 @@ public record AttendanceReportSummary(
         String projectId,
         String period,
         int workingDays,
-        int presentDays,
+        double presentDays,
         int halfDays,
         int leaveDays,
         int absentDays,
